--- airbyte-integrations/connectors/source-kustomer-singer/.gitignore
@@ -1,1 +1,0 @@
-NEW_SOURCE_CHECKLIST.md
