--- airbyte-integrations/connectors/source-kustomer-singer/integration_tests/abnormal_state.json
@@ -1,13 +1,0 @@
-{
-  "bookmarks": {
-    "conversations": "2050-12-26T18:35:41.583Z",
-    "customers": "2050-12-26T18:39:38.186Z",
-    "kobjects": "2050-12-26T18:24:18.283Z",
-    "messages": "2050-12-26T18:39:45.018Z",
-    "notes": "2050-12-26T18:39:45.018Z",
-    "shortcuts": "2050-12-25T18:00:00Z",
-    "tags": "2050-12-25T18:00:00Z",
-    "teams": "2050-12-25T18:00:00Z",
-    "users": "2050-12-25T18:00:00Z"
-  }
-}
