--- airbyte-integrations/connectors/source-kustomer-singer/integration_tests/sample_state.json
@@ -1,13 +1,0 @@
-{
-  "bookmarks": {
-    "conversations": "2021-12-25T18:00:00Z",
-    "customers": "2021-12-25T18:00:00Z",
-    "kobjects": "2021-12-25T18:00:00Z",
-    "messages": "2021-12-25T18:00:00Z",
-    "notes": "2021-12-25T18:00:00Z",
-    "shortcuts": "2021-12-25T18:00:00Z",
-    "tags": "2021-12-25T18:00:00Z",
-    "teams": "2021-12-25T18:00:00Z",
-    "users": "2021-12-25T18:00:00Z"
-  }
-}
