--- airbyte-integrations/connectors/source-kustomer-singer/metadata.yaml
@@ -1,24 +1,0 @@
-data:
-  connectorSubtype: api
-  connectorType: source
-  definitionId: cd06e646-31bf-4dc8-af48-cbc6530fcad3
-  dockerImageTag: 0.1.2
-  dockerRepository: airbyte/source-kustomer-singer
-  githubIssueLabel: source-kustomer-singer
-  icon: kustomer.svg
-  license: MIT
-  name: Kustomer
-  registries:
-    cloud:
-      enabled: true
-    oss:
-      enabled: true
-  releaseStage: alpha
-  documentationUrl: https://docs.airbyte.com/integrations/sources/kustomer-singer
-  tags:
-    - language:python
-  ab_internal:
-    sl: 100
-    ql: 100
-  supportLevel: community
-metadataSpecVersion: "1.0"
