--- airbyte-integrations/connectors/source-kustomer-singer/requirements.txt
@@ -1,2 +1,0 @@
-# This file is autogenerated -- only edit if you know what you are doing. Use setup.py for declaring dependencies.
-# -e .
