+++ airbyte-integrations/connectors/source-kustomer/__init__.py
@@ -1,7 +1,3 @@
 #
 # Copyright (c) 2023 Airbyte, Inc., all rights reserved.
 #
-
-
-def test_example_method():
-    assert True
