+++ airbyte-integrations/connectors/source-kustomer/acceptance-test-config.yml
@@ -1,0 +1,39 @@
+# See [Connector Acceptance Tests](https://docs.airbyte.com/connector-development/testing-connectors/connector-acceptance-tests-reference)
+# for more information about how to configure these tests
+connector_image: airbyte/source-kustomer:dev
+acceptance_tests:
+  spec:
+    tests:
+      - spec_path: "source_kustomer/spec.yaml"
+  connection:
+    tests:
+      - config_path: "secrets/config.json"
+        status: "succeed"
+      - config_path: "integration_tests/invalid_config.json"
+        status: "failed"
+  discovery:
+    tests:
+      - config_path: "secrets/config.json"
+  basic_read:
+    tests:
+      - config_path: "secrets/config.json"
+        configured_catalog_path: "integration_tests/configured_catalog.json"
+        empty_streams: []
+# TODO uncomment this block to specify that the tests should assert the connector outputs the records provided in the input file a file
+#        expect_records:
+#          path: "integration_tests/expected_records.jsonl"
+#          extra_fields: no
+#          exact_order: no
+#          extra_records: yes
+  incremental: 
+    bypass_reason: "This connector does not implement incremental sync"
+# TODO uncomment this block this block if your connector implements incremental sync: 
+#    tests:
+#      - config_path: "secrets/config.json"
+#        configured_catalog_path: "integration_tests/configured_catalog.json"
+#        future_state:
+#          future_state_path: "integration_tests/abnormal_state.json"
+  full_refresh:
+    tests:
+      - config_path: "secrets/config.json"
+        configured_catalog_path: "integration_tests/configured_catalog.json"
