+++ airbyte-integrations/connectors/source-kustomer/acceptance-test-docker.sh
@@ -1,2 +1,3 @@
 #!/usr/bin/env sh
+
 source "$(git rev-parse --show-toplevel)/airbyte-integrations/bases/connector-acceptance-test/acceptance-test-docker.sh"
