+++ airbyte-integrations/connectors/source-kustomer/integration_tests/__init__.py
@@ -1,0 +1,3 @@
+#
+# Copyright (c) 2023 Airbyte, Inc., all rights reserved.
+#
