+++ airbyte-integrations/connectors/source-kustomer/integration_tests/invalid_config.json
