+++ airbyte-integrations/connectors/source-kustomer/integration_tests/sample_config.json
@@ -1,0 +1,4 @@
+{
+  "api_token": "xxxxxxxxxx",
+  "start_date": "2023-10-10T00:00:00Z"
+}
