+++ airbyte-integrations/connectors/source-kustomer/main.py
@@ -2,11 +2,12 @@
 # Copyright (c) 2023 Airbyte, Inc., all rights reserved.
 #
 
+
 import sys
 
 from airbyte_cdk.entrypoint import launch
-from source_kustomer_singer import SourceKustomerSinger
+from source_kustomer import SourceKustomer
 
 if __name__ == "__main__":
-    source = SourceKustomerSinger()
+    source = SourceKustomer()
     launch(source, sys.argv[1:])
