+++ airbyte-integrations/connectors/source-kustomer/requirements.txt
@@ -1,0 +1,2 @@
+-e ../../bases/connector-acceptance-test
+-e .
