+++ airbyte-integrations/connectors/source-kustomer/source_kustomer/__init__.py
@@ -1,0 +1,8 @@
+#
+# Copyright (c) 2023 Airbyte, Inc., all rights reserved.
+#
+
+
+from .source import SourceKustomer
+
+__all__ = ["SourceKustomer"]
