+++ airbyte-integrations/connectors/source-kustomer/source_kustomer/manifest.yaml
@@ -1,0 +1,171 @@
+version: "0.29.0"
+
+definitions:
+  selector:
+    type: RecordSelector
+    extractor:
+      type: DpathExtractor
+      field_path: ["data"]
+
+  requester:
+    type: HttpRequester
+    url_base: "https://api.kustomerapp.com/v1"
+    http_method: "GET"
+    authenticator:
+      type: BearerAuthenticator
+      api_token: "{{ config['api_token'] }}"
+
+  base_paginator:
+    type: "DefaultPaginator"
+    page_size_option:
+      type: "RequestOption"
+      inject_into: "request_parameter"
+      field_name: "pageSize"
+    pagination_strategy:
+      type: "PageIncrement"
+      page_size: 10
+    page_token_option:
+      type: "RequestOption"
+      inject_into: "request_parameter"
+      field_name: "page"
+
+  retriever:
+    type: SimpleRetriever
+    record_selector:
+      $ref: "#/definitions/selector"
+    paginator:
+      $ref: "#/definitions/base_paginator"
+    requester:
+      $ref: "#/definitions/requester"
+
+  base_stream:
+    type: DeclarativeStream
+    retriever:
+      $ref: "#/definitions/retriever"
+
+  incremental_sync_base:
+    type: DatetimeBasedCursor
+    cursor_field: "{{ parameters.incremental_cursor }}"
+    datetime_format: "%Y-%m-%dT%H:%M:%S.%fZ"
+    cursor_granularity: "PT0.000001S"
+    lookback_window: "P31D"
+    start_datetime:
+      datetime: "{{ config['start_date'] }}"
+      datetime_format: "%Y-%m-%dT%H:%M:%SZ"
+    end_datetime:
+      datetime: "{{ today_utc() }}"
+      datetime_format: "%Y-%m-%d"
+    step: "P1M"
+
+  customers_stream:
+    $ref: "#/definitions/base_stream"
+    $parameters:
+      name: "customers"
+      primary_key: "id"
+      incremental_cursor: "updated_at"
+      path: "/customers"
+    incremental_sync:
+      $ref: "#/definitions/incremental_sync_base"
+
+  conversations_stream:
+    $ref: "#/definitions/base_stream"
+    $parameters:
+      name: "conversations"
+      primary_key: "id"
+      incremental_cursor: "updated_at"
+      path: "/conversations"
+    incremental_sync:
+      $ref: "#/definitions/incremental_sync_base"
+
+  kobjects_stream:
+    $ref: "#/definitions/base_stream"
+    $parameters:
+      name: "kobjects"
+      primary_key: "id"
+      incremental_cursor: "updated_at"
+      path: "/kobjects"
+    incremental_sync:
+      $ref: "#/definitions/incremental_sync_base"
+
+  messages_stream:
+    $ref: "#/definitions/base_stream"
+    $parameters:
+      name: "messages"
+      primary_key: "id"
+      incremental_cursor: "updated_at"
+      path: "/messages"
+    incremental_sync:
+      $ref: "#/definitions/incremental_sync_base"
+
+  notes_stream:
+    $ref: "#/definitions/base_stream"
+    $parameters:
+      name: "notes"
+      primary_key: "id"
+      incremental_cursor: "updated_at"
+      path: "/notes"
+    incremental_sync:
+      $ref: "#/definitions/incremental_sync_base"
+
+  shortcuts_stream:
+    $ref: "#/definitions/base_stream"
+    $parameters:
+      name: "shortcuts"
+      primary_key: "id"
+      incremental_cursor: "updated_at"
+      path: "/shortcuts"
+    incremental_sync:
+      $ref: "#/definitions/incremental_sync_base"
+
+  tags_stream:
+    $ref: "#/definitions/base_stream"
+    $parameters:
+      name: "tags"
+      primary_key: "id"
+      incremental_cursor: "updated_at"
+      path: "/tags"
+    incremental_sync:
+      $ref: "#/definitions/incremental_sync_base"
+
+  teams_stream:
+    $ref: "#/definitions/base_stream"
+    $parameters:
+      name: "teams"
+      primary_key: "id"
+      incremental_cursor: "updated_at"
+      path: "/teams"
+
+  users_stream:
+    $ref: "#/definitions/base_stream"
+    $parameters:
+      name: "users"
+      primary_key: "id"
+      incremental_cursor: "updated_at"
+      path: "/users"
+    incremental_sync:
+      $ref: "#/definitions/incremental_sync_base"
+
+streams:
+  - "#/definitions/customers_stream"
+  - "#/definitions/conversations_stream"
+  - "#/definitions/kobjects_stream"
+  - "#/definitions/messages_stream"
+  - "#/definitions/notes_stream"
+  - "#/definitions/shortcuts_stream"
+  - "#/definitions/tags_stream"
+  - "#/definitions/teams_stream"
+  - "#/definitions/users_stream"
+
+check:
+  type: CheckStream
+  stream_names:
+    - "customers"
+    - "conversations"
+    - "kobjects"
+    - "messages"
+    - "notes"
+    - "shortcuts"
+    - "tags"
+    - "teams"
+    - "users"
+
