+++ airbyte-integrations/connectors/source-kustomer/source_kustomer/schemas/notes.json
@@ -1,0 +1,92 @@
+{
+  "$schema": "http://json-schema.org/draft-07/schema#",
+  "type": "object",
+  "additionalProperties": true,
+  "properties": {
+    "body": { "type": ["null", "string"] },
+    "conversation": {
+      "properties": {
+        "id": { "type": ["null", "string"] },
+        "type": { "type": ["null", "string"] }
+      },
+      "type": ["null", "object"],
+      "additionalProperties": false
+    },
+    "created_at": { "format": "date-time", "type": ["null", "string"] },
+    "created_by": {
+      "properties": {
+        "id": { "type": ["null", "string"] },
+        "type": { "type": ["null", "string"] }
+      },
+      "type": ["null", "object"],
+      "additionalProperties": false
+    },
+    "customer": {
+      "properties": {
+        "id": { "type": ["null", "string"] },
+        "type": { "type": ["null", "string"] }
+      },
+      "type": ["null", "object"],
+      "additionalProperties": false
+    },
+    "deleted": { "type": ["null", "boolean"] },
+    "deleted_at": { "format": "date-time", "type": ["null", "string"] },
+    "deleted_by": {
+      "properties": {
+        "id": { "type": ["null", "string"] },
+        "type": { "type": ["null", "string"] }
+      },
+      "type": ["null", "object"],
+      "additionalProperties": false
+    },
+    "external_id": { "type": ["null", "string"] },
+    "id": { "type": ["null", "string"] },
+    "imported_at": {
+      "format": "date-time",
+      "type": ["null", "string"]
+    },
+    "links": {
+      "properties": { "self": { "type": ["null", "string"] } },
+      "type": ["null", "object"],
+      "additionalProperties": false
+    },
+    "modified_at": {
+      "format": "date-time",
+      "type": ["null", "string"]
+    },
+    "modified_by": {
+      "properties": {
+        "id": { "type": ["null", "string"] },
+        "type": { "type": ["null", "string"] }
+      },
+      "type": ["null", "object"],
+      "additionalProperties": false
+    },
+    "org": {
+      "properties": {
+        "id": { "type": ["null", "string"] },
+        "type": { "type": ["null", "string"] }
+      },
+      "type": ["null", "object"],
+      "additionalProperties": false
+    },
+    "type": { "type": ["null", "string"] },
+    "updated_at": { "format": "date-time", "type": ["null", "string"] },
+    "user_mentions": {
+      "anyOf": [
+        {
+          "type": "array",
+          "items": {
+            "type": ["null", "object"],
+            "additionalProperties": false,
+            "properties": {
+              "id": { "type": ["null", "string"] },
+              "type": { "type": ["null", "string"] }
+            }
+          }
+        },
+        { "type": "null" }
+      ]
+    }
+  }
+}
