+++ airbyte-integrations/connectors/source-kustomer/source_kustomer/schemas/users.json
@@ -1,0 +1,75 @@
+{
+  "$schema": "http://json-schema.org/draft-07/schema#",
+  "type": "object",
+  "additionalProperties": true,
+  "properties": {
+    "avatar_url": { "type": ["null", "string"] },
+    "created_at": { "format": "date-time", "type": ["null", "string"] },
+    "created_by": {
+      "properties": {
+        "id": { "type": ["null", "string"] },
+        "type": { "type": ["null", "string"] }
+      },
+      "type": ["null", "object"],
+      "additionalProperties": false
+    },
+    "deleted_at": { "type": ["null", "string"] },
+    "display_name": { "type": ["null", "string"] },
+    "email": { "type": ["null", "string"] },
+    "email_signature": { "type": ["null", "string"] },
+    "email_verified_at": { "type": ["null", "string"] },
+    "first_email_verified_at": { "type": ["null", "string"] },
+    "id": { "type": ["null", "string"] },
+    "links": {
+      "properties": { "self": { "type": ["null", "string"] } },
+      "type": ["null", "object"],
+      "additionalProperties": false
+    },
+    "mobile": { "type": ["null", "string"] },
+    "modified_at": { "type": ["null", "string"] },
+    "modified_by": {
+      "properties": {
+        "id": { "type": ["null", "string"] },
+        "type": { "type": ["null", "string"] }
+      },
+      "type": ["null", "object"],
+      "additionalProperties": false
+    },
+    "name": { "type": ["null", "string"] },
+    "org": {
+      "properties": {
+        "id": { "type": ["null", "string"] },
+        "type": { "type": ["null", "string"] }
+      },
+      "type": ["null", "object"],
+      "additionalProperties": false
+    },
+    "password": {
+      "properties": {
+        "allow_new": { "type": "boolean" },
+        "force_new": { "type": "boolean" },
+        "updated_at": {
+          "format": "date-time",
+          "type": ["null", "string"]
+        }
+      },
+      "type": ["null", "object"],
+      "additionalProperties": false
+    },
+    "role_groups": {
+      "anyOf": [
+        { "items": { "type": ["null", "string"] }, "type": "array" },
+        { "type": "null" }
+      ]
+    },
+    "roles": {
+      "anyOf": [
+        { "items": { "type": ["null", "string"] }, "type": "array" },
+        { "type": "null" }
+      ]
+    },
+    "type": { "type": ["null", "string"] },
+    "updated_at": { "format": "date-time", "type": ["null", "string"] },
+    "user_type": { "type": ["null", "string"] }
+  }
+}
