+++ airbyte-integrations/connectors/source-kustomer/source_kustomer/spec.yaml
@@ -1,0 +1,22 @@
+documentationUrl: https://docs.airbyte.com/integrations/sources/kustomer
+connectionSpecification:
+  $schema: http://json-schema.org/draft-07/schema#
+  title: Source Kustomer Singer Spec
+  type: object
+  required:
+  - api_token
+  - start_date
+  additionalProperties: true
+  properties:
+    api_token:
+      title: API Token
+      type: string
+      description: Kustomer API Token. See the <a href="https://developer.kustomer.com/kustomer-api-docs/reference/authentication">docs</a>
+        on how to obtain this
+      airbyte_secret: true
+    start_date:
+      title: Start Date
+      type: string
+      description: The date from which you'd like to replicate the data
+      examples:
+      - '2019-01-01T00:00:00Z'
